requirements.txt
azure-ai-formrecognizer
pypdf2 
reportlab 
streamlit 
openai 
python-dotenv